pyproject.toml
[project]
name = "ai-company"
version = "0.1.0"
description = "AI 에이전트로 구성된 회사 - Agentic AI 오케스트레이션 시스템"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AI Company Team"}
]

dependencies = [
    # Core - Orchestration
    "langgraph>=0.2.61",
    "langchain>=0.3.14",
    "langchain-anthropic>=0.3.1",
    "langchain-core>=0.3.29",

    # Memory & Storage
    "chromadb>=0.5.23",
    "sqlalchemy>=2.0.36",

    # Utilities
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.1",

    # Async support
    "asyncio>=3.4.3",
    "aiofiles>=24.1.0",

    # Web Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "langsmith>=0.2.10",
    "ruff>=0.8.6",
    "mypy>=1.14.1",
]

[project.scripts]
ai-company = "src.cli:cli"
ai-company-server = "uvicorn:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
